fix: fix high sev implicit narrowing conversion alert given by CodeQL

+++ orders-service/src/main/java/com/orders_service/Order.java
@@ -335,7 +335,7 @@ public class Order {
   }
 
   public boolean validate_total_check() {
-    float sum = 0.0f;
+    double sum = 0.0f;
     for (OrderItem item : this.orderItems) {
       sum += item.getSubtotal();
     }
